Refactored JdbcResolver and SpringJdbcResolver into a common super class
They are almost identical. Also more documentation for subclassing
as well as a test case for a custom subclass.
