+++ flyway-core/src/test/java/org/flywaydb/core/FlywaySmallTest.java
@@ -16,11 +16,14 @@
 package org.flywaydb.core;
 
 import org.flywaydb.core.api.FlywayException;
+import org.flywaydb.core.api.MigrationInfo;
+import org.flywaydb.core.api.MigrationInfoService;
 import org.flywaydb.core.api.callback.FlywayCallback;
 import org.flywaydb.core.api.resolver.MigrationResolver;
 import org.flywaydb.core.internal.dbsupport.DbSupport;
 import org.flywaydb.core.internal.dbsupport.Schema;
 import org.flywaydb.core.internal.metadatatable.MetaDataTable;
+import org.flywaydb.core.internal.resolver.CustomSqlMigrationResolver;
 import org.flywaydb.core.internal.resolver.MyConfigurationAwareCustomMigrationResolver;
 import org.flywaydb.core.internal.resolver.MyCustomMigrationResolver;
 import org.flywaydb.core.internal.util.jdbc.DriverDataSource;
@@ -28,6 +31,8 @@ import org.junit.Test;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Properties;
 
 import static org.junit.Assert.*;
@@ -222,4 +227,25 @@ public class FlywaySmallTest {
 
         configResolver.assertFlywayConfigurationIsSet();
     }
+
+    @Test
+    public void replaceResolverWithSubclass() {
+        Flyway flyway = new Flyway();
+        flyway.setDataSource("jdbc:h2:mem:flyway_test;DB_CLOSE_DELAY=-1", "sa", "");
+        flyway.setResolvers(new CustomSqlMigrationResolver());
+        flyway.setLocations("migration/subdir");
+        flyway.setSkipDefaultResolvers(true);
+
+        flyway.clean(); // classloader problems
+        MigrationInfoService info = flyway.info();
+
+        List<MigrationInfo> migrationInfos = Arrays.asList(info.all());
+
+        assertEquals(3, migrationInfos.size());
+        assertEquals("99.1", migrationInfos.get(0).getVersion().getVersion());
+        assertEquals("99.1.1", migrationInfos.get(1).getVersion().getVersion());
+        assertEquals("99.2.0", migrationInfos.get(2).getVersion().getVersion());
+    }
+
+
 }
